requirements.txt
flask==3.0.0
flask-cors==4.0.0
yt-dlp==2024.12.13
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
instabot==0.117.0
requests==2.31.0
pillow==10.1.0
moviepy==1.0.3
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
ffmpeg-python==0.2.0
pytube==15.0.0
selenium==4.16.0
webdriver-manager==4.0.1
opencv-python==4.8.1.78
numpy==1.24.3
python-telegram-bot==20.7
schedule==1.2.0
psutil==5.9.6
tqdm==4.66.1 